separated main function into index.js, add handleCreateUser
