+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "jest",
     "test-watch": "jest --watch",
-    "start": "node net.js"
+    "start": "node index.js"
   },
   "repository": {
     "type": "git",
